worker.py moved to appropriate place
